Allow `/` as shortcut key in Firefox, prevent arrow shortcuts from triggering browser behaviors
/ in firefox is a Find in Page shortcut, which revoked focus from the
page. And at least for me, Arrows + cmd + shift was triggering scroll
behaviors in a Tree Tab extension I run (tabs as a list to the side).
Since that and other behaviors are hard to predict, I think it's best to
always preventDefault() once it's clear that a known shortcut is going
to be triggered.

Update tracker with latest zzfx from cb07fa9ca